refactoring movie list event to separate file
