+++ lib/domain/blocs/movie_list_event.dart
@@ -1,0 +1,15 @@
+abstract class MovieListEvent {}
+
+class MovieListEventLoadNextPage extends MovieListEvent {
+  final String locale;
+
+  MovieListEventLoadNextPage(this.locale);
+}
+
+class MovieListEventLoadReset extends MovieListEvent {}
+
+class MovieListEventSearchMovie extends MovieListEvent {
+  final String query;
+
+  MovieListEventSearchMovie(this.query);
+}
